requirements.txt
tensorflow-gpu==2.2.0
opencv-python==4.1.2.30
bcolz==1.2.1
sklearn2
PyYAML
tqdm
